Token Auth and Templating

+++ Todo-List/web/app.py
@@ -1,11 +1,14 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request, jsonify, render_template, flash
 from flask_sqlalchemy import SQLAlchemy
 import enum
 import datetime
+import random
+import string
 import config
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = config.SQLALCHEMY_DB_URI
+app.config['SECRET_KEY'] = config.SECRET_KEY
 
 db = SQLAlchemy(app)
 
@@ -17,6 +20,7 @@ class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(20), unique=True, nullable=False)
     password = db.Column(db.Text, nullable=False)
+    token = db.Column(db.String(48), unique=True, nullable=False)
 
     def __repr__(self):
         return f'<User "{self.username}">'
@@ -33,18 +37,25 @@ class Task(db.Model):
     def __repr__(self):
         return f'<Task "{self.title}">'
 
+random_str = lambda N: ''.join(
+    random.SystemRandom().choice(string.ascii_uppercase + string.ascii_lowercase + string.digits) for _ in range(N))
+
+
+@app.route('/')
+def index():
+    return render_template('index.html')
 
 @app.route('/tasks/add', methods=['POST'])
 def add_task():
     try:
+        token = request.form['token']
         title = request.form['task_title']
         desc = request.form['task_desc']
         status = request.form['task_status']
         start_date = request.form['task_start_date']
         end_date = request.form['task_end_date']
-        user_id = int(request.form['user_id'])
 
-        user = User.query.get(user_id)
+        user = User.query.filter_by(token=token).first()
         if user:
             start_date_datetime = datetime.datetime.strptime(
                 start_date, config.DATETIME_FORMAT)
@@ -52,7 +63,7 @@ def add_task():
             end_date_datetime = datetime.datetime.strptime(
                 end_date, config.DATETIME_FORMAT)
             
-            task = Task(title=title, desc=desc, status=status, start_date=start_date_datetime, end_date=end_date_datetime, user_id=user_id)
+            task = Task(title=title, desc=desc, status=status, start_date=start_date_datetime, end_date=end_date_datetime, user_id=user.id)
             db.session.add(task)
             db.session.commit()
 
@@ -76,8 +87,10 @@ def add_task():
 @app.route('/q/tasks/', methods=['POST'])
 def get_tasks():
     try:
+        token = request.form['token']
         username = request.form['username']
-        user = User.query.filter_by(username=username).first()
+        user = User.query.filter_by(username=username, token=token).first()
+        
         if user:
             user_tasks = Task.query.filter_by(user_id=user.id)
             tasks_list = []
@@ -113,6 +126,7 @@ def get_tasks():
 @app.route('/tasks/edit', methods=['POST'])
 def edit_task():
     try:
+        token = request.form['token']
         title = request.form['task_title']
         desc = request.form['task_desc']
         status = request.form['task_status']
@@ -120,25 +134,31 @@ def edit_task():
         end_date = request.form['task_end_date']
         task_id = request.form['task_id']
 
-        task = Task.query.get(task_id)
-        if task:
-            start_date_datetime = datetime.datetime.strptime(
-                start_date, config.DATETIME_FORMAT)
-            
-            end_date_datetime = datetime.datetime.strptime(
-                end_date, config.DATETIME_FORMAT)
-            
-            task.title = title
-            task.desc = desc
-            task.status = status
-            task.start_date = start_date_datetime
-            task.end_date = end_date_datetime
-            
-            db.session.commit()
+        user = User.query.filter_by(token=token)
+        if user:
+            task = Task.query.get(task_id)
+            if task:
+                start_date_datetime = datetime.datetime.strptime(
+                    start_date, config.DATETIME_FORMAT)
+                
+                end_date_datetime = datetime.datetime.strptime(
+                    end_date, config.DATETIME_FORMAT)
+                
+                task.title = title
+                task.desc = desc
+                task.status = status
+                task.start_date = start_date_datetime
+                task.end_date = end_date_datetime
+                
+                db.session.commit()
 
-            data = {
-                'status': 'ok'
-            }
+                data = {
+                    'status': 'ok'
+                }
+            else:
+                data = {
+                    'status': 'not found'
+                }
         else:
             data = {
                 'status': 'not found'
@@ -153,16 +173,23 @@ def edit_task():
 @app.route('/tasks/remove', methods=['POST'])
 def remove_task():
     try:
-        task_id = request.form['task_id'] # TODO: in error handling handle this with key_error
+        token = request.form['token']
+        task_id = request.form['task_id']
 
-        task = Task.query.get(task_id)
-        if task:
-            db.session.delete(task)
-            db.session.commit()
+        user = User.query.filter_by(token=token)
+        if user:
+            task = Task.query.get(task_id)
+            if task:
+                db.session.delete(task)
+                db.session.commit()
 
-            data = {
-                'status': 'ok'
-            }
+                data = {
+                    'status': 'ok'
+                }
+            else:
+                data = {
+                    'status': 'not found'
+                }
         else:
             data = {
                 'status': 'not found'
@@ -174,25 +201,25 @@ def remove_task():
         }
     return jsonify(data)
 
-@app.route('/account/register', methods=['POST'])
+@app.route('/account/register', methods=['GET', 'POST'])
 def register():
     try:
-        username = request.form['username'] # TODO: Handle When a User Exists and unique constraint failed 
-        password = request.form['password']
-        user = User(username=username, password=password)
-        db.session.add(user)
-        db.session.commit()
+        if request.method == 'POST':
+            username = request.form['username']
+            password = request.form['password']
 
-        data = {
-            'status': 'ok'
-        }
+            if User.query.filter_by(username=username).first():
+                flash('این نام کاربری قبلا ثبت شده')
+            else:
+                token = random_str(48)
+                user = User(username=username, password=password, token=token)
+                db.session.add(user)
+                db.session.commit()
+                flash(f'عملیات با موفقیت انجام شد. توکن شما:  {token}')
     except Exception as e:
-        data = {
-            'status': 'error',
-            'exception': str(e)
-        }
-
-    return jsonify(data)
+        flash('خطایی رخ داد. بعدا دوباره امتحان کنید')
+    
+    return render_template('register.html')
 
 
 @app.route('/account/login', methods=['POST'])
@@ -203,7 +230,8 @@ def login():
         user = User.query.filter_by(username=username, password=password).first()
         if user:
             data = {
-                'status': 'ok'
+                'status': 'ok',
+                'token': user.token
             }
         else:
             data = {
